Add router gateway support
When a gateway is set on router, the plugin starts a NetNS source NAT
service instance between the a dedicated private network and the public.
Then it creates a routing table with a default route that point to the
left interface of the NetNS service instance private port.
Each time a network is attached with aqn interface to the virtual
router, the route table is applied.

Change-Id: I53592099bda87628e90a53159e92f99777ccff10
Signed-off-by: Édouard Thuleau <edouard.thuleau@cloudwatt.com>