Currently, if we receive a delete of a property from irond, the ifmap_server_table.cc code will clear the bit for the property but not remove the data that the SetProperty() had copied in.
Now, consider the case where I add property P1, then add P2 and finally remove
P2. Removing P2 is actually changing the node and this change must be
downloaded to the agent. But, our current CalculateCrc() code will return the
same value since the data for P2 has not been removed.

Fix is for CalculateCrc() to use a field only if IsPropertySet() is true.

Change-Id: I0f98179d0150eefcad8a80a6f87a8345577acaf3