Fix id-perms in cassandra client generator
Before this fix, the generator generated this
piece of code. So with a access by key on a potentially
None object.

field = obj_dict.get('id_perms', None)
field['created'] = datetime.datetime.utcnow().isoformat()
field['last_modified'] = field['created']
if field is not None:
    self._create_prop(bch, obj_ids['uuid'], 'id_perms', field)

This moves the access to the created and last attributes only
if the field is a dict, as following:

field = obj_dict.get('id_perms', None)
if field is not None:
    self._create_prop(bch, obj_ids['uuid'], 'id_perms', field)
    field['created'] = datetime.datetime.utcnow().isoformat()
    field['last_modified'] = field['created']

Change-Id: Id24ff052b45fa778d326e5e4f3fdd2fd8b55e7d7