Route aggregate support in schema transformer
Added support for route aggregate objects in schema transformer. Schema
transformer will watch for route aggregate objects, and when they are linked to
a service instance, it will also link it to the corresponding left or right
routing instance.

Also made route aggregate a child of project similar to routing policy.

Change-Id: I22a4def2ed0508b392b8df0156068ed55f0f2d1c
Partial-Bug: 1500698