Adding port-security-enabled parameter to the schema
This is a neutron extension that we claim to support, but it is
not currently implemented. To impement this, we will need to
support this flag at network and port level. At the port level,
if port security is disabled, no security groups can be applied
to the port and the port can communicate with all other ports.

At the time of port creation, if a specific value is not provided,
we will copy if from the corresponding network. The default value
for both will be True, i.e., port security is enabled unless
explicitly disabled by the user.

Change-Id: Ibaf659ee9b383268fad4b95f09adf9f90b3a87e6
Related-Bug: 1632961