Implement fate sharing for SIs in a service chain
If either the left or right interface of any SI in a chain is down, stop
re-orignating all routes for all SIs in the chain.  This prevents black
holing of traffic when an intermediate SI in a chain fails.

Introduce the notion of a service chain group to define a collection of
SIs that are part of a given service chain. It is assumed that the group
name/uuid will be managed by the schema transformer.

The control node maintains a list of RoutingInstances for a group (each
RI represents half of a SI). It updates this list when the configuration
changes.  Further, it tracks the state of the connected route for each
RI. If the connected route for any of the RIs in a service chain group
is down, the operational status of the group is set to down and all the
re-originated routes for the RIs in the group are deleted. Conversely,
when the connected routes for all RIs in a group are up, the operational
status of the group is changed to up and all re-originated routes are
added/updated.

Add a LifetimeRef from the ServiceChainMgr<T> to the BgpServer to ensure
that a ServiceChainMgr<T> cannot be destroyed till all the groups in the
ServiceChainMgr<T> have been deleted. Note that the deletion of RIs is
still triggered via the RoutingInstanceMgr.

Change-Id: Id190b0046c1b5cfdc2d03f8d6f90a38557719376
Partial-Bug: 1736208