Making VNC API pip installable
VNC API is much dependent on the config common code right now. Also we are
hard coding the version of the generated package. Which makes the package
static and dependent on other code. Hence, to upload to pip as a package,
we need to make it independent of the other modules and also automate the
versioning of the package.

Change-Id: I211cdead591e6312787aae34b0da37fe0432ad95
Partial-Bug: #1711447