BGP schema: Support default tunnel encap on peer
As part of adding support for inter AS Option C, there is a need to be able to
mainipulate the tunnel encapsulation.
BGP Schema changes are made to configure the Default Tunnel Encapsulation List
per address family, per peer. Note, that this list is not expected to be
configured for non-VPN address families including labeled inet (earlier commit
for supporting this feature). Support for the labeled inet family would require
a minor change, i.e., not blindly strip communities as it does now.
Added 'mpls' as an encapsulation supported in the unicast XMPP schema, have NOT
done the same for the multicast and Ethernet XMPP schemas at this time.
Note that currently any string can be sent on the wire, so this appears to be
for documentation only,  however is not up to date.
Partial-Bug: #1743517
Change-Id: I315b7290acbb72adea98546fb821ea4f738dd0be