[schema] Set draft switch property under Global System Config
In the first implementation of the draft mode for security resources, we
set the property switch under the project resource for the project scope
case and under the global policy management for global scope case. This
is not compliant which what we do habitually, a deployment config
property should be a property of the Global System Config resource.

Change-Id: Id1ef5769ae09f7a109f76362fa4bd2293f0626ba
Partial-Bug: #1759501