[Config] Introduce a new property at LR to indicate the type of LR.
In Contrail, we have been unable to create any LR for SNAT purpose
if VXLAN routing is enabled and vice versa. This schema change is to
facilitate such creations.

The idea is to move the vxlan_routing flag from a project level
to per LR basis. What this means is, now we can create a LR for SNAT
and also create another LR for VXLAN routing.

This change is only for implementing the schema changes

Change-Id: I5cc96707bf1860e2865ea781c30ab252ac8dace2
Partial-Jira-Bug: JCB-192482