Remove reference from virtual-network to logical-router
The reference from a virtual-network to a logical-router is not used
anymore. It was introduced in the early Contrail release (before 2.10)
to allocate a VN/RI per LR but it was replaced by an allocated RT
per LR to leak routes to connected VN [1]. But recently, a patch
re-introduced it but it's still not used.

[1] https://github.com/Juniper/contrail-controller/commit/ebbc4f484cf4f3d7b915e0d54fd1d657ee8a167c
[2] https://github.com/Juniper/contrail-api-client/commit/3b00dff14d82cc17b9cab5f04ccaba0accaaa186

Closes-Jira-Bug: CEM-5125

Change-Id: Id119398325aee5c6930ba17edb3d5d79a8080704