Make generateDS imap-frontend runnable as per documentation.
This is in preparation of the next couple changes.
generateDS can be run from the command line to parse the .xsd
schema files and generate a variety of outputs, selected with
the -g option.

Contrail documentation is genearted by using -g ifmap-frontend,
which also generates python files.

According to the help screen of generateDS, and content of SConscript
files, a command like:

  ../../generateDS.py -f -o foo -g ifmap-frontend ../all_cfg.xsd

would generate a set of files like foo_client.py, foo_common.py, foo_test.py,
... all based on the all_cfg.xsd schema, and would also generate the
contrail_openapi.{rst,json,adoc} and the heat/ directory used by
sphinx to generate documentation.

Before this change:
- The command line above crashed, with a cryptic message.

After some investigation, it turns out that the -o option has to match
some hardcoded values, and an environment variable compatible with
those parameters has to be passed in for the generation to work.

In this change:
- Stop requiring an undocumented environment variable to deduce the
  base of the output path. The output path is passed with the -o variable.
- Stop appending a hardcoded path to decide the output path. Again,
  this is controlled by the -o option, and the paths passed in.
- Stop assuming that the output is called 'resource'. The name of the
  output is controlled by the -o parameter.

Note that the change is transparent / neutral to the rest of the
build. The computed path now deduced from -o look to be the same that
would be deduced with the hardcoded values.

To test this change:
- run the script via scons before and after printing out the heat_path
  variable.  Before and after the change, the computed value is the same, no
  difference.
- the scons log before and after is exactly the same, except for
  name of tempfiles, timestamps, and time to run tests.
- ignoring zip files, tar.gz, the build/ directory of a build without
  the change and a build with the change leads exactly the same output.

Further:
- grep of the entire codebase shows no use of a variable called
  HEAT_BUILDTOP.
- there is only one use of -g ifmap-frontend I could find with
  a simple grep.

Which probably means it's ok to deprecate the parameter, and if it
works in the one place I tested, it should be fine.

Change-Id: I566b612e2a7938db2cc83baee27d5d674afc1087