[Schema] Including Fabric reference to Logical Router object
In order to check if a VN is associated to more than one LR in the same
fabric, we need to track LR and fabric associations. Although at present
we can track this using LR --> PR --> Fabric link, it is too much of
an indirection and may lead to performance issues esp. on UI.

Therefore adding this reference on the LR object.

Change-Id: Ie649c194eb4cc508c15374bf1aee03f35f687a98
closes-jira-bug: CEM-12445
target-release: R2008