CEM-24418: Add description field to security group rule structure
Description: PolicyRuleType is the structure describing a security
group rule. If the user creates a security group rule and adds a
description to the respective rule via OpenStack API, that description
will not be stored because the field is missing from Contrail.

Root cause: Missing optional field from PolicyRuleType

Solution: Add description field to PolicyRuleType

Closes-jira-bug: CEM-24418
Signed-off-by: Anda-Maria Nicolae <mnicolae@juniper.net>
Change-Id: I5133643e4e7bbc52a278fb4c526c834b4ff28db7